+++ tests/test_SA203.py
@@ -7,6 +7,8 @@ from src.flake8_sqlalchemy2 import Checker
 
 def test_import_simple():
     code = """\
+from typing import List
+
 from sqlalchemy import Integer, ForeignKey
 from sqlalchemy.orm import (
     Mapped,
@@ -25,7 +27,7 @@ class Parent(Base):
     __tablename__ = "parent"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    children: DynamicMapped["Child"] = relationship(back_populates="parent")
+    children: Mapped[List["Child"]] = relationship(backref="parent")
 
 
 class Child(Base):
@@ -33,21 +35,21 @@ class Child(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     parent_id: Mapped[int] = mapped_column(ForeignKey("parent.id"))
-    parent: Mapped["Parent"] = relationship(back_populates="children")
 """
 
     tree = ast.parse(code)
     plugin = Checker(tree)
     assert [f"{line}:{col + 1} {msg}" for line, col, msg, _ in plugin.run()] == [
-        "19:15 SA202 Use of legacy collection `DynamicMapped` consider using `WriteOnlyMapped` instead"
+        "21:52 SA203 Use of legacy relationship `backref` consider using `back_populates` instead"
     ]
 
 
 def test_import_orm():
     code = """\
+from typing import List
+
 import sqlalchemy as sa
 from sqlalchemy import orm
-
 
 class Base(orm.DeclarativeBase):
     pass
@@ -57,7 +59,7 @@ class Parent(Base):
     __tablename__ = "parent"
 
     id: orm.Mapped[int] = orm.mapped_column(primary_key=True)
-    children: orm.DynamicMapped["Child"] = orm.relationship(back_populates="parent")
+    children: orm.Mapped[List["Child"]] = orm.relationship(backref="parent")
 
 
 class Child(Base):
@@ -65,11 +67,10 @@ class Child(Base):
 
     id: orm.Mapped[int] = orm.mapped_column(primary_key=True)
     parent_id: orm.Mapped[int] = orm.mapped_column(sa.ForeignKey("parent.id"))
-    parent: orm.Mapped["Parent"] = orm.relationship(back_populates="children")
 """
 
     tree = ast.parse(code)
     plugin = Checker(tree)
     assert [f"{line}:{col + 1} {msg}" for line, col, msg, _ in plugin.run()] == [
-        "13:15 SA202 Use of legacy collection `DynamicMapped` consider using `WriteOnlyMapped` instead"
+        "14:60 SA203 Use of legacy relationship `backref` consider using `back_populates` instead"
     ]